refactor to navigation
