+++ example/package.json
@@ -10,6 +10,7 @@
     "build:ios": "react-native build-ios --scheme HoneycombOpentelemetryReactNativeExample --mode Debug --extra-params \"-sdk iphonesimulator CC=clang CPLUSPLUS=clang++ LD=clang LDPLUSPLUS=clang++ GCC_OPTIMIZATION_LEVEL=0 GCC_PRECOMPILE_PREFIX_HEADER=YES ASSETCATALOG_COMPILER_OPTIMIZATION=time DEBUG_INFORMATION_FORMAT=dwarf COMPILER_INDEX_STORE_ENABLE=NO\""
   },
   "dependencies": {
+    "@react-navigation/native-stack": "^7.3.13",
     "react": "19.0.0",
     "react-native": "0.78.1",
     "react-native-safe-area-context": "^5.4.1",
